// datahub-web-react/src/types.generated.ts
export type CorpGroup = Entity & {
    __typename?: 'CorpGroup';
    /** The primary key of the group */
    urn: Scalars['String'];
    /** A standard Entity Type */
    type: EntityType;
    /** Group name eg wherehows dev, ask_metadata */
    name: Scalars['String'];
    /** Ownership metadata of the Corp Group */
    ownership?: Maybe<Ownership>;
    /** Additional read only properties about the group */
    properties?: Maybe<CorpGroupProperties>;
    /** Additional read write properties about the group */
    editableProperties?: Maybe<CorpGroupEditableProperties>;
    /** Granular API for querying edges extending from this entity */
    relationships?: Maybe<EntityRelationshipsResult>;
    /**
     * Deprecated, use properties field instead
     * Additional read only info about the group
     * @deprecated Field no longer supported
     */
    info?: Maybe<CorpGroupInfo>;
};
// ...
export type CorpUser = Entity & {
    __typename?: 'CorpUser';
    /** The primary key of the user */
    urn: Scalars['String'];
    /** The standard Entity Type */
    type: EntityType;
    /**
     * A username associated with the user
     * This uniquely identifies the user within DataHub
     */
    username: Scalars['String'];
    /** Additional read only properties about the corp user */
    properties?: Maybe<CorpUserProperties>;
    /** Read write properties about the corp user */
    editableProperties?: Maybe<CorpUserEditableProperties>;
    /** The status of the user */
    status?: Maybe<CorpUserStatus>;
    /** The tags associated with the user */
    tags?: Maybe<GlobalTags>;
    /** Granular API for querying edges extending from this entity */
    relationships?: Maybe<EntityRelationshipsResult>;
    /**
     * Deprecated, use properties field instead
     * Additional read only info about the corp user
     * @deprecated Field no longer supported
     */
    info?: Maybe<CorpUserInfo>;
    /**
     * Deprecated, use editableProperties field instead
     * Read write info about the corp user
     * @deprecated Field no longer supported
     */
    editableInfo?: Maybe<CorpUserEditableInfo>;
    /**
     * Deprecated, use the tags field instead
     * The structured tags associated with the user
     * @deprecated Field no longer supported
     */
    globalTags?: Maybe<GlobalTags>;
};
// ...
export type Dashboard = EntityWithRelationships &
    Entity & {
        __typename?: 'Dashboard';
        /** The primary key of the Dashboard */
        urn: Scalars['String'];
        /** A standard Entity Type */
        type: EntityType;
        /** The parent container in which the entity resides */
        container?: Maybe<Container>;
        /** Recursively get the lineage of containers for this entity */
        parentContainers?: Maybe<ParentContainersResult>;
        /**
         * The dashboard tool name
         * Note that this will soon be deprecated in favor of a standardized notion of Data Platform
         */
        tool: Scalars['String'];
        /** An id unique within the dashboard tool */
        dashboardId: Scalars['String'];
        /** Additional read only properties about the dashboard */
        properties?: Maybe<DashboardProperties>;
        /** Additional read write properties about the dashboard */
        editableProperties?: Maybe<DashboardEditableProperties>;
        /** Ownership metadata of the dashboard */
        ownership?: Maybe<Ownership>;
        /** Status metadata of the dashboard */
        status?: Maybe<Status>;
        /** The deprecation status of the dashboard */
        deprecation?: Maybe<Deprecation>;
        /** The tags associated with the dashboard */
        tags?: Maybe<GlobalTags>;
        /** References to internal resources related to the dashboard */
        institutionalMemory?: Maybe<InstitutionalMemory>;
        /** The structured glossary terms associated with the dashboard */
        glossaryTerms?: Maybe<GlossaryTerms>;
        /** The Domain associated with the Dashboard */
        domain?: Maybe<Domain>;
        /** The specific instance of the data platform that this entity belongs to */
        dataPlatformInstance?: Maybe<DataPlatformInstance>;
        /** Granular API for querying edges extending from this entity */
        relationships?: Maybe<EntityRelationshipsResult>;
        /** Edges extending from this entity grouped by direction in the lineage graph */
        lineage?: Maybe<EntityLineageResult>;
        /**
         * Deprecated, use properties field instead
         * Additional read only information about the dashboard
         * @deprecated Field no longer supported
         */
        info?: Maybe<DashboardInfo>;
        /**
         * Deprecated, use editableProperties instead
         * Additional read write properties about the Dashboard
         * @deprecated Field no longer supported
         */
        editableInfo?: Maybe<DashboardEditableProperties>;
        /**
         * Deprecated, use tags field instead
         * The structured tags associated with the dashboard
         * @deprecated Field no longer supported
         */
        globalTags?: Maybe<GlobalTags>;
        /** Standardized platform urn where the dashboard is defined */
        platform: DataPlatform;
    };
// ...
export type DataPlatform = Entity & {
    __typename?: 'DataPlatform';
    /** Urn of the data platform */
    urn: Scalars['String'];
    /** A standard Entity Type */
    type: EntityType;
    /** Name of the data platform */
    name: Scalars['String'];
    /** Additional read only properties associated with a data platform */
    properties?: Maybe<DataPlatformProperties>;
    /**
     * Deprecated, use properties displayName instead
     * Display name of the data platform
     * @deprecated Field no longer supported
     */
    displayName?: Maybe<Scalars['String']>;
    /**
     * Deprecated, use properties field instead
     * Additional properties associated with a data platform
     * @deprecated Field no longer supported
     */
    info?: Maybe<DataPlatformInfo>;
    /** Edges extending from this entity */
    relationships?: Maybe<EntityRelationshipsResult>;
};
// ...
export type Domain = Entity & {
    __typename?: 'Domain';
    /** The primary key of the domain */
    urn: Scalars['String'];
    /** A standard Entity Type */
    type: EntityType;
    /** Id of the domain */
    id: Scalars['String'];
    /** Properties about a domain */
    properties?: Maybe<DomainProperties>;
    /** Ownership metadata of the dataset */
    ownership?: Maybe<Ownership>;
    /** References to internal resources related to the dataset */
    institutionalMemory?: Maybe<InstitutionalMemory>;
    /** Children entities inside of the Domain */
    entities?: Maybe<SearchResults>;
    /** Edges extending from this entity */
    relationships?: Maybe<EntityRelationshipsResult>;
};
// ...
export enum EntityType {
    /** A Domain containing Metadata Entities */
    Domain = 'DOMAIN',
    /** The Dataset Entity */
    Dataset = 'DATASET',
    /** The CorpUser Entity */
    CorpUser = 'CORP_USER',
    /** The CorpGroup Entity */
    CorpGroup = 'CORP_GROUP',
    /** The DataPlatform Entity */
    DataPlatform = 'DATA_PLATFORM',
    /** The Dashboard Entity */
    Dashboard = 'DASHBOARD',
    /** The Notebook Entity */
    Notebook = 'NOTEBOOK',
    /** The Chart Entity */
    Chart = 'CHART',
    /** The Data Flow (or Data Pipeline) Entity, */
    DataFlow = 'DATA_FLOW',
    /** The Data Job (or Data Task) Entity */
    DataJob = 'DATA_JOB',
    /** The Tag Entity */
    Tag = 'TAG',
    /** The Glossary Term Entity */
    GlossaryTerm = 'GLOSSARY_TERM',
    /** The Glossary Node Entity */
    GlossaryNode = 'GLOSSARY_NODE',
    /** A container of Metadata Entities */
    Container = 'CONTAINER',
    /** The ML Model Entity */
    Mlmodel = 'MLMODEL',
    /** The MLModelGroup Entity */
    MlmodelGroup = 'MLMODEL_GROUP',
    /** ML Feature Table Entity */
    MlfeatureTable = 'MLFEATURE_TABLE',
    /** The ML Feature Entity */
    Mlfeature = 'MLFEATURE',
    /** The ML Primary Key Entity */
    MlprimaryKey = 'MLPRIMARY_KEY',
    /** A DataHub Managed Ingestion Source */
    IngestionSource = 'INGESTION_SOURCE',
    /** A DataHub ExecutionRequest */
    ExecutionRequest = 'EXECUTION_REQUEST',
    /** A DataHub Assertion */
    Assertion = 'ASSERTION',
    /** An instance of an individual run of a data job or data flow */
    DataProcessInstance = 'DATA_PROCESS_INSTANCE',
    /** Data Platform Instance Entity */
    DataPlatformInstance = 'DATA_PLATFORM_INSTANCE',
    /** A DataHub Access Token */
    AccessToken = 'ACCESS_TOKEN',
    /** A DataHub Test */
    Test = 'TEST',
}
// ...
export type GlossaryNode = Entity & {
    __typename?: 'GlossaryNode';
    /** The primary key of the glossary term */
    urn: Scalars['String'];
    /** Ownership metadata of the glossary term */
    ownership?: Maybe<Ownership>;
    /** A standard Entity Type */
    type: EntityType;
    /** Additional properties associated with the Glossary Term */
    properties?: Maybe<GlossaryNodeProperties>;
    /** Edges extending from this entity */
    relationships?: Maybe<EntityRelationshipsResult>;
    /** Recursively get the lineage of glossary nodes for this entity */
    parentNodes?: Maybe<ParentNodesResult>;
};
// ...
export type Notebook = Entity & {
    __typename?: 'Notebook';
    /** The primary key of the Notebook */
    urn: Scalars['String'];
    /** A standard Entity Type */
    type: EntityType;
    /** The Notebook tool name */
    tool: Scalars['String'];
    /** An id unique within the Notebook tool */
    notebookId: Scalars['String'];
    /** Additional read only information about the Notebook */
    info?: Maybe<NotebookInfo>;
    /** Additional read write properties about the Notebook */
    editableProperties?: Maybe<NotebookEditableProperties>;
    /** Ownership metadata of the Notebook */
    ownership?: Maybe<Ownership>;
    /** Status metadata of the Notebook */
    status?: Maybe<Status>;
    /** The content of this Notebook */
    content: NotebookContent;
    /** The tags associated with the Notebook */
    tags?: Maybe<GlobalTags>;
    /** References to internal resources related to the Notebook */
    institutionalMemory?: Maybe<InstitutionalMemory>;
    /** The Domain associated with the Notebook */
    domain?: Maybe<Domain>;
    /** The specific instance of the data platform that this entity belongs to */
    dataPlatformInstance?: Maybe<DataPlatformInstance>;
    /** Edges extending from this entity */
    relationships?: Maybe<EntityRelationshipsResult>;
    /** Sub Types that this entity implements */
    subTypes?: Maybe<SubTypes>;
    /** The structured glossary terms associated with the notebook */
    glossaryTerms?: Maybe<GlossaryTerms>;
    /** Standardized platform. */
    platform: DataPlatform;
};
// ...
export type PlatformPrivileges = {
    __typename?: 'PlatformPrivileges';
    /** Whether the user should be able to view analytics */
    viewAnalytics: Scalars['Boolean'];
    /** Whether the user should be able to manage policies */
    managePolicies: Scalars['Boolean'];
    /** Whether the user should be able to manage users & groups */
    manageIdentities: Scalars['Boolean'];
    /** Whether the user should be able to generate personal access tokens */
    generatePersonalAccessTokens: Scalars['Boolean'];
    /** Whether the user should be able to manage Domains */
    manageDomains: Scalars['Boolean'];
    /** Whether the user is able to manage UI-based ingestion */
    manageIngestion: Scalars['Boolean'];
    /** Whether the user is able to manage UI-based secrets */
    manageSecrets: Scalars['Boolean'];
    /** Whether the user should be able to manage tokens on behalf of other users. */
    manageTokens: Scalars['Boolean'];
    /** Whether the user is able to manage Tests */
    manageTests: Scalars['Boolean'];
    /** Whether the user should be able to manage Glossaries */
    manageGlossaries: Scalars['Boolean'];
};